requirements.txt
fastapi>=0.95.0 
pydantic[email]>=1.10.7,<2.0.0
uvicorn>=0.21.1 
gunicorn==20.1.0
psycopg2-binary==2.9.3
asyncpg==0.26.0 
SQLAlchemy>=1.4.40
SQLAlchemy-Utils>=0.38.3
alembic>=1.8.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.5
python-dotenv==0.21.0
pytz>=2023.3 
Jinja2>=3.1.2
#sqlmodel>=0.0.8